# Dispatch analyze to segments