@baobao0206 2020-05-13
I have a performance test on the skew hash
```
create table t1(a int, b int);
create table t2(a int, b int);
insert into t1 select i, i from generate_series(1, 1000000)i;
insert into t2 select i, i from generate_series(1, 1000000)i;
insert into t2 select i%100, i%100 from generate_series(1, 5000000)i;
```
```
demo=# select count(*) from t1, t2 where t1.a=t2.a;
  count
---------
 5950000
(1 row)

Time: 3612.739 ms
```
Query use 3612 ms without skew hash

```
demo=# select count(*) from t1, t2 where t1.a=t2.a;
  count
---------
 5950000
(1 row)

Time: 2524.882 ms
```
Query use 2534ms with skew hash

They have same plan.