@hlinnaka 2020-05-13
With this patch, we'll do twice as much work on ANALYZE. First we call `analyze_rel()' which collects the sample rows, and computes the statistics based on it. Then we collect a sample again, and send that sample to the caller. Can we avoid the duplication?