@hlinnaka 2020-05-13
Perhaps modify analyze_rel() function so that it returns the sample rows to the caller, in addition computing the stats?