@d 2020-05-13
> Query use 2534ms with skew hash
> 
> They have same plan.

Can you detail the steps it takes to see the performance difference? Do you have a patch locally?