@baobao0206 2020-05-14
@d Just run the querys with and without the patch.
```
create table t1(a int, b int);
create table t2(a int, b int);
insert into t1 select i, i from generate_series(1, 1000000)i;
insert into t2 select i, i from generate_series(1, 1000000)i;
insert into t2 select i%100, i%100 from generate_series(1, 5000000)i;
analyze t1;
analyze t2;
```