@baobao0206 2020-05-14
@hlinnaka Have a new patch on the code and avoid repeated data collection