@d 2020-05-14
> Just run the querys with and without the patch.

Stupid question: if I run `ANALYZE t1; ANALYZE t2;` in utility mode on every segment -- in addition to running it from master -- will that show a difference?