@baobao0206 2020-05-15
> in addition to running it from master

I think the only difference is that it only sampled tuples once in this patch, and sampled tuples twice using the method you said