@d 2020-05-21
I couldn't see the speed up with this patch applied, I'm wondering whether there's a way to tell if I did anything stupid, or whether the skew hash execution actually happened...