@d 2020-05-21
I ran a `perf` against both before and after the patch, and the skew hash portion of `nodeHash` doesn't seem to show up in the trace in either execution of your example. Are you sure this is the example to demonstrate that skew hash execution happened at all (besides having a benefit)?