@baobao0206 2020-05-21
@d We have not used Perf for performance analysis. But this patch does get a stable performance improvement in my example. I will make a careful performance analysis tomorrow.