@d 2020-05-21
> @d We have not used Perf for performance analysis. But this patch does get a stable performance improvement in my example. I will make a careful performance analysis tomorrow.

To be clear, `perf` is not necessary actually. I was too lazy to attach a debugger so I use `perf` as a lazy man's `gdb` to see any outsized portion of execution. I see an almost identical profile for before and after the patch.

To get a stable profile, here's what I did:
1. I quit Zoom;
1. quit Chrome; and
1. quit all my IDE's, then;
1. switched my laptop CPU from `powersave` to `performance`
1. ran perf against the query