@hlinnaka 2020-05-22
Note that the skew bucket optimization only kicks in if the hash table is large enough to require multiple batches. I had to `set statement_mem='2MB'` to see the effect with @baobao0206's test query.