@d 2020-05-22
> Note that the skew bucket optimization only kicks in if the hash table is large enough to require multiple batches. I had to `set statement_mem='2MB'`

Oops, how could I have forgotten that... @melanieplageman and I looked into this before (commit d3ccff0434e4d005aefcaa8c6b9f7b0f770718c2 née #9257 ) and we even got a spun-out PR https://github.com/greenplum-db/gpdb/pull/9630

Anyway, an `EXPLAIN ANALYZE` comparing the execution with and without segment stats shows the difference:

Without segment stats:

```
                                                             QUERY PLAN                                                              
-------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate (actual time=1530.299..1530.300 rows=1 loops=1)
   Output: count(*)
   ->  Gather Motion 3:1  (slice1; segments: 3) (actual time=1156.318..1530.292 rows=3 loops=1)
         Output: (PARTIAL count(*))
         ->  Partial Aggregate (actual time=1529.547..1529.547 rows=1 loops=1)
               Output: PARTIAL count(*)
               ->  Hash Join (actual time=286.939..1399.647 rows=2233167 loops=1)
                     Hash Cond: (t2.a = t1.a)
                     work_mem: 3301kB  Segments: 3  Max: 1102kB (segment 2)  Workfile: (3 spilling)
                     Work_mem wanted: 7941K bytes avg, 7958K bytes max (seg2) to lessen workfile I/O affecting 3 workers.
                     Extra Text: (seg0)   Initial batch 0:
 (seg0)     Wrote 3420K bytes to inner workfile.
 (seg0)     Wrote 37927K bytes to outer workfile.
 (seg0)   Initial batches 1..7:
 (seg0)     Read 3420K bytes from inner workfile: 489K avg x 7 nonempty batches, 493K max.
 (seg0)     Read 37927K bytes from outer workfile: 5419K avg x 7 nonempty batches, 10579K max.
 (seg0)   Hash chain length 2.8 avg, 13 max, using 120702 of 131072 buckets.
                     Extra Text: (seg2)   Initial batch 0:
 (seg2)     Wrote 3428K bytes to inner workfile.
 (seg2)     Wrote 25244K bytes to outer workfile.
 (seg2)   Initial batches 1..7:
 (seg2)     Read 3428K bytes from inner workfile: 490K avg x 7 nonempty batches, 494K max.
 (seg2)     Read 25244K bytes from outer workfile: 3607K avg x 7 nonempty batches, 5695K max.
 (seg2)   Hash chain length 2.8 avg, 12 max, using 120858 of 131072 buckets.Initial batch 0:
 
                     ->  Seq Scan on public.t2 (actual time=0.056..138.370 rows=2233167 loops=1)
                           Output: t2.a, t2.b
                     ->  Hash (actual time=287.851..287.851 rows=334042 loops=1)
                           Output: t1.a
                           ->  Seq Scan on public.t1 (actual time=0.027..129.349 rows=334042 loops=1)
                                 Output: t1.a
 Planning time: 0.285 ms
   (slice0)    Executor memory: 53K bytes.
 * (slice1)    Executor memory: 1609K bytes avg x 3 workers, 1609K bytes max (seg0).  Work_mem: 1102K bytes max, 7958K bytes wanted.
 Memory used:  2048kB
 Memory wanted:  8557kB
 Optimizer: Postgres query optimizer
 Settings: optimizer=off
 Execution time: 1535.039 ms
(39 rows)
```

With this patch applied:
```
                                                             QUERY PLAN                                                              
-------------------------------------------------------------------------------------------------------------------------------------
 Finalize Aggregate (actual time=1329.389..1329.389 rows=1 loops=1)
   Output: count(*)
   ->  Gather Motion 3:1  (slice1; segments: 3) (actual time=1158.947..1329.381 rows=3 loops=1)
         Output: (PARTIAL count(*))
         ->  Partial Aggregate (actual time=1313.236..1313.236 rows=1 loops=1)
               Output: PARTIAL count(*)
               ->  Hash Join (actual time=292.565..1186.701 rows=2233167 loops=1)
                     Hash Cond: (t2.a = t1.a)
                     work_mem: 3310kB  Segments: 3  Max: 1106kB (segment 1)  Workfile: (3 spilling)
                     Work_mem wanted: 7943K bytes avg, 7959K bytes max (seg2) to lessen workfile I/O affecting 3 workers.
                     Extra Text: (seg0)   Initial batch 0:
 (seg0)     Wrote 3420K bytes to inner workfile.
 (seg0)     Wrote 5699K bytes to outer workfile.
 (seg0)   Initial batches 1..7:
 (seg0)     Read 3420K bytes from inner workfile: 489K avg x 7 nonempty batches, 493K max.
 (seg0)     Read 5699K bytes from outer workfile: 815K avg x 7 nonempty batches, 821K max.
 (seg0)   Hash chain length 2.8 avg, 13 max, using 120697 of 131072 buckets.
                     Extra Text: (seg2)   Initial batch 0:
 (seg2)     Wrote 3428K bytes to inner workfile.
 (seg2)     Wrote 5713K bytes to outer workfile.
 (seg2)   Initial batches 1..7:
 (seg2)     Read 3428K bytes from inner workfile: 490K avg x 7 nonempty batches, 494K max.
 (seg2)     Read 5713K bytes from outer workfile: 817K avg x 7 nonempty batches, 823K max.
 (seg2)   Hash chain length 2.8 avg, 12 max, using 120855 of 131072 buckets.Initial batch 0:
 
                     ->  Seq Scan on public.t2 (actual time=0.240..367.402 rows=2233167 loops=1)
                           Output: t2.a, t2.b
                     ->  Hash (actual time=358.817..358.817 rows=334042 loops=1)
                           Output: t1.a
                           ->  Seq Scan on public.t1 (actual time=0.023..202.438 rows=334042 loops=1)
                                 Output: t1.a
 Planning time: 0.430 ms
   (slice0)    Executor memory: 53K bytes.
 * (slice1)    Executor memory: 1613K bytes avg x 3 workers, 1613K bytes max (seg0).  Work_mem: 1106K bytes max, 7959K bytes wanted.
 Memory used:  2048kB
 Memory wanted:  8558kB
 Optimizer: Postgres query optimizer
 Settings: optimizer=off
 Execution time: 1333.979 ms
(39 rows)
```

And I noticed the decrease on the maximum outer batch size.